refactor: separate the clothing items types from the types file
